Laboratory Exercise #5: Loops in C
A C program that uses loops and conditional statements to control what values are added. The program must do the following:
Ask the user to enter numbers one by one.
If the user enters a positive number, it will be added to the total sum.
If the user enters zero or a negative number, the program will stop asking for input.
After stopping, the program will display the total sum of all positive numbers entered.